# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is sample dapp to provide trials of LBD service",
        "title": "Link Cinema API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "0.1"
    },
    "basePath": "/api/v0",
    "paths": {
        "/test/transaction": {
            "get": {
                "description": "Retrieve a Transaction using its hash",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "test"
                ],
                "summary": "Get a transaction",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Transaction hash used for searching",
                        "name": "txhash",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Transaction with the provided hash",
                        "schema": {
                            "$ref": "#/definitions/service.Transaction"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "service.Amount": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "integer"
                },
                "denom": {
                    "type": "string"
                }
            }
        },
        "service.Attribute": {
            "type": "object",
            "properties": {
                "key": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "service.Event": {
            "type": "object",
            "properties": {
                "attributes": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Attribute"
                    }
                },
                "type": {
                    "type": "string"
                }
            }
        },
        "service.Fee": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Amount"
                    }
                },
                "gas": {
                    "type": "integer"
                }
            }
        },
...
        "service.Log": {
            "type": "object",
            "properties": {
                "events": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Event"
                    }
                },
                "log": {
                    "type": "string"
                },
                "msg_index": {
                    "type": "integer"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "service.Message": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {}
            }
        },
...
        "service.PubKey": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "type": "string"
                }
            }
        },
...
        "service.Signature": {
            "type": "object",
            "properties": {
                "pubKey": {
                    "$ref": "#/definitions/service.PubKey"
                },
                "signature": {
                    "type": "string"
                }
            }
        },
...
        "service.Transaction": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "gasUsed": {
                    "type": "integer"
                },
                "gasWanted": {
                    "type": "integer"
                },
                "height": {
                    "type": "integer"
                },
                "index": {
                    "type": "integer"
                },
                "logs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Log"
                    }
                },
                "raw_log": {
                    "type": "string"
                },
                "timestamp": {
                    "type": "integer"
                },
                "tx": {
                    "$ref": "#/definitions/service.Tx"
                },
                "txhash": {
                    "type": "string"
                }
            }
        },
...
        "service.Tx": {
            "type": "object",
            "properties": {
                "type": {
                    "type": "string"
                },
                "value": {
                    "$ref": "#/definitions/service.TxValue"
                }
            }
        },
        "service.TxValue": {
            "type": "object",
            "properties": {
                "fee": {
                    "$ref": "#/definitions/service.Fee"
                },
                "memo": {
                    "type": "string"
                },
                "msg": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Message"
                    }
                },
                "signatures": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.Signature"
                    }
                }
            }
        },
...
    }
}